refactor: 边框13 tsx 改造
