+++ packages/docs/docs/Border/BorderBox13/demo.vue
@@ -1,9 +1,12 @@
 <template>
   <div demo-bg>
-    <dv-border-box-13>
+    <dv-border-box13>
       <div dv-bg>
         dv-border-box-13
       </div>
-    </dv-border-box-13>
+    </dv-border-box13>
   </div>
 </template>
+<script setup lang="ts">
+import { BorderBox13 as DvBorderBox13 } from '@kjgl77/datav-vue3'
+</script>
